Split AssetBuilder into two classes
AssetTool now contains all the argument parsing code. It is pretty much
the same as the old AssetBuilder except without any code to convert
assets. The new AssetBuilder class has all of that, and is now all
static code and instances, like AssetWriter.

This is still quite messy, but the two halves can now be refactored
independently.