Remove working_path from packer
Adjusting every path supplied by the yaml or the command line is problematic.
The paths in the yaml are strings so we can't tell them apart from other
options easily. That means the yaml parser has to know the details of every
builder's unique arguments.

Instead of doing that, we can just change the current working directory, ie
cd to the directory containing the config file. This should be completely
equivalent, and is one line of code.